requirements.txt
Flask==2.3.3
Flask_Cors==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
Requests==2.31.0